@@ -35,7 +35,7 @@
     </organization>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <file.encoding>UTF-8</file.encoding>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
@@ -47,20 +47,20 @@
         <project.build.resourceEncoding>${file.encoding}</project.build.resourceEncoding>
         <project.reporting.outputEncoding>${file.encoding}</project.reporting.outputEncoding>
 
-        <spring-boot.version>2.7.18</spring-boot.version>
+        <spring-boot.version>3.3.1</spring-boot.version>
 
         <c3p0.version>0.9.5.5</c3p0.version>
-        <druid.version>1.2.23</druid.version>
-        <plexus-spring-util.version>${spring-boot.version}</plexus-spring-util.version>
+        <druid.version>1.2.21</druid.version>
+        <plexus-spring-util.version>3.3.1.1</plexus-spring-util.version>
         <vibur-dbcp.version>25.0</vibur-dbcp.version>
 
-        <javadoc-docent-maven-plugin.version>1.8.2</javadoc-docent-maven-plugin.version>
-        <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
-        <maven-install-plugin.version>2.5.2</maven-install-plugin.version>
-        <maven-jar-plugin.version>3.2.2</maven-jar-plugin.version>
-        <maven-javadoc-plugin.version>3.6.2</maven-javadoc-plugin.version>
-        <maven-resources-plugin.version>3.2.0</maven-resources-plugin.version>
-        <maven-source-plugin.version>3.2.1</maven-source-plugin.version>
+        <javadoc-docent-maven-plugin.version>2.0.0</javadoc-docent-maven-plugin.version>
+        <maven-compiler-plugin.version>3.13.0</maven-compiler-plugin.version>
+        <maven-install-plugin.version>3.1.2</maven-install-plugin.version>
+        <maven-jar-plugin.version>3.4.2</maven-jar-plugin.version>
+        <maven-javadoc-plugin.version>3.6.3</maven-javadoc-plugin.version>
+        <maven-resources-plugin.version>3.3.1</maven-resources-plugin.version>
+        <maven-source-plugin.version>3.3.1</maven-source-plugin.version>
         <maven-gpg-plugin.version>3.1.0</maven-gpg-plugin.version>
         <central-publishing-maven-plugin.version>0.5.0</central-publishing-maven-plugin.version>
     </properties>
